Refactored repo.js with function imported from merge.js
